opt: performance improvements for histograms
This commit improves the performance of histograms in the optimizer
by avoiding allocating and copying the full histogram unless strictly
necessary. Additionally, it changes the code for filtering histograms
to use binary search instead of performing a linear scan.

Release note: None

+++ pkg/sql/opt/cat/table.go
@@ -182,15 +182,17 @@ type TableStatistic interface {
 	Histogram() []HistogramBucket
 }
 
-// HistogramBucket contains the data for a single histogram bucket.
+// HistogramBucket contains the data for a single histogram bucket. Note
+// that NumEq, NumRange, and DistinctRange are floats so the statisticsBuilder
+// can apply filters to the histogram.
 type HistogramBucket struct {
 	// NumEq is the estimated number of values equal to UpperBound.
-	NumEq uint64
+	NumEq float64
 
 	// NumRange is the estimated number of values between the upper bound of the
 	// previous bucket and UpperBound (both boundaries are exclusive).
 	// The first bucket should always have NumRange=0.
-	NumRange uint64
+	NumRange float64
 
 	// DistinctRange is the estimated number of distinct values between the upper
 	// bound of the previous bucket and UpperBound (both boundaries are
